Rename bert2lstm
