+++ KD_Lib/BERT2LSTM/__init__.py
@@ -1,2 +1,2 @@
-from .bert2lstm import Bert2LSTM
+from .bert2lstm import BERT2LSTM
 from .utils import get_essentials
