+++ KD_Lib/BERT2LSTM/bert2lstm.py
@@ -6,14 +6,14 @@ import torch.nn.functional as F
 from transformers import BertForSequenceClassification, AdamW, BertTokenizer
 
 from KD_Lib.common import BaseClass
-from KD_Lib.Bert2Lstm.utils import df_to_dataset, batch_to_inputs, set_seed
+from KD_Lib.BERT2LSTM.utils import df_to_dataset, batch_to_inputs, set_seed
 
 import numpy as np
 import matplotlib.pyplot as plt
 from copy import deepcopy
 
 
-class Bert2LSTM(BaseClass):
+class BERT2LSTM(BaseClass):
     """
     Implementation of Knowledge distillation from the paper "Distilling Task-Specific
     Knowledge from BERT into Simple Neural Networks" https://arxiv.org/pdf/1903.12136.pdf
